requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
pillow==10.0.0
scikit-learn==1.3.1